+++ src/Kernels/Distance/Hamming.php
@@ -1,0 +1,53 @@
+<?php
+
+namespace Rubix\ML\Kernels\Distance;
+
+use Rubix\ML\Other\Helpers\DataType;
+
+/**
+ * Hamming
+ *
+ * A categorical distance function that measures distance as the number of
+ * substitutions necessary to convert one sample to the other.
+ *
+ * References:
+ * [1] R. W. Hamming. (1950). Error detecting and error correcting codes.
+ *
+ * @category    Machine Learning
+ * @package     Rubix/ML
+ * @author      Andrew DalPino
+ */
+class Hamming implements Distance
+{
+    /**
+     * Return the data types that this kernel is compatible with.
+     *
+     * @return int[]
+     */
+    public function compatibility() : array
+    {
+        return [
+            DataType::CATEGORICAL,
+        ];
+    }
+
+    /**
+     * Compute the distance between two vectors.
+     *
+     * @param array $a
+     * @param array $b
+     * @return float
+     */
+    public function compute(array $a, array $b) : float
+    {
+        $distance = 0;
+
+        foreach ($a as $i => $value) {
+            if ($value !== $b[$i]) {
+                $distance++;
+            }
+        }
+
+        return (float) $distance;
+    }
+}